Adapt `Cone` constructor to match type of keyword arguments
When no type is explicitely passed to the `Cone` constructor, it should take the promoted type only from the keyword arguments. Thus, the default parameter are defined as `Int`. If all parameters are `Int`, the `Cone` methods that perform the type promotion ensure that the `Cone` is always of type `Float`.