Adapt `Box` constructor to match type of keyword arguments
When no type is explicitely passed to the `Box` constructor, it should take the promoted type only from the keyword arguments. Thus, the default parameter are defined as `Int`. If all parameters are `Int`, the `Box` methods that perform the type promotion ensure that the `Box` is always of type `Float`.