Fix plotting of electric field lines
When plotting electric field lines, charges are drifting according to
the electric field. The initial spawn positions are determined by
sampling the contacts in steps of `sampling` and moving them into the
bulk of the detector by an `offset`.

In `ConstructiveSolidGeometry`, the surfaces of a `VolumePrimitive` are
defined such that their `normal` vector always points towards the center
of the `VolumePrimitive`. In order to compute the correct spawn
positions, the charges need to be spawned outside of the contact, i.e.
in negative direction of the `normal` vector.

This behavior does not appear if the contacts have a thickness of 0. It
does, however, when the contacts have non-zero width.