Use `G4SingleParticleSource` instead of `G4ParticleGun` for isotope source generation.
The resulted energy spectrum is not a Dirac delta function anymore, but
it is not similar to a typical 207-Bi spectrum either. Maybe it's caused
by the following line?

`SetParticleMomentumDirection(GetAngDist(gun), data.direction)`

It does not make sense to me that one specifies the direction of the
momentum of an isotope source. I modified it only to make the code
compile.